Fixed the "duplicate" Shop instances!!! :))
The shop now contains the same items in both the views. At first, a new list of items was being generated by the program, because the List & Grid view scripts derived from ShopView, making a new instance of ShopView for both the views, thus also making a new ShopModel, ShopController (and therefore a new Inventory). Now, there is only one instance of ShopView because a Singleton was implemented.